build 1.25

+++ TqkLibrary.Scrcpy.Wpf/ControlChain.cs
@@ -74,7 +74,7 @@ namespace TqkLibrary.Scrcpy.Wpf
         public bool InjectKeycode(AndroidKeyEventAction action, AndroidKeyCode keycode, uint repeat = 1, AndroidKeyEventMeta metaState = AndroidKeyEventMeta.META_NONE)
             => this.FixAll(x => x.InjectKeycode(action, keycode, repeat, metaState));
 
-        public bool InjectScrollEvent(Rectangle position, int vScroll, int hScroll = 0, AndroidMotionEventButton button = AndroidMotionEventButton.BUTTON_PRIMARY)
+        public bool InjectScrollEvent(Rectangle position, float vScroll, float hScroll = 0, AndroidMotionEventButton button = AndroidMotionEventButton.BUTTON_PRIMARY)
             => this.FixAll(x => x.InjectScrollEvent(new Rectangle(position.Location, x.Scrcpy.ScreenSize), vScroll, hScroll, button));
 
         public bool InjectText(string text)
